Simplify JSON Schema headers
