+++ src/core/jsonschema/include/sourcemeta/core/jsonschema_frame.h
@@ -5,19 +5,22 @@
 #include <sourcemeta/core/jsonschema_export.h>
 #endif
 
+#include <sourcemeta/core/json.h>
 #include <sourcemeta/core/jsonpointer.h>
-#include <sourcemeta/core/jsonschema_keywords.h>
+
 #include <sourcemeta/core/jsonschema_resolver.h>
+#include <sourcemeta/core/jsonschema_types.h>
 #include <sourcemeta/core/jsonschema_walker.h>
 
-#include <cstdint>    // std::uint8_t
-#include <functional> // std::reference_wrapper
-#include <map>        // std::map
-#include <optional>   // std::optional
-#include <string>     // std::string
-#include <tuple>      // std::tuple
-#include <utility>    // std::pair
-#include <vector>     // std::vector
+#include <cstdint>       // std::uint8_t
+#include <functional>    // std::reference_wrapper
+#include <map>           // std::map
+#include <optional>      // std::optional
+#include <string>        // std::string
+#include <tuple>         // std::tuple
+#include <unordered_set> // std::set
+#include <utility>       // std::pair
+#include <vector>        // std::vector
 
 namespace sourcemeta::core {
 
@@ -204,6 +207,59 @@ private:
 #pragma warning(default : 4251 4275)
 #endif
 };
+
+// TODO: Eventually generalize this to detecting cross-keyword dependencies as
+// part of framing
+
+/// @ingroup jsonschema
+struct SchemaUnevaluatedEntry {
+  /// The absolute pointers of the static keyword dependencies
+  std::set<Pointer> static_dependencies;
+  /// The absolute pointers of the static keyword dependencies
+  std::set<Pointer> dynamic_dependencies;
+  /// Whether the entry cannot be fully resolved, which means
+  /// there might be unknown dynamic dependencies
+  bool unresolved{false};
+};
+
+/// @ingroup jsonschema
+/// The flattened set of unevaluated cases in the schema by absolute URI
+using SchemaUnevaluatedEntries = std::map<std::string, SchemaUnevaluatedEntry>;
+
+/// @ingroup jsonschema
+///
+/// This function performs a static analysis pass on `unevaluatedProperties` and
+/// `unevaluatedItems` occurences throughout the entire schema (if any).
+///
+/// For example:
+///
+/// ```cpp
+/// #include <sourcemeta/core/json.h>
+/// #include <sourcemeta/core/jsonschema.h>
+/// #include <cassert>
+///
+/// const sourcemeta::core::JSON document =
+///     sourcemeta::core::parse_json(R"JSON({
+///   "$schema": "https://json-schema.org/draft/2020-12/schema",
+///   "unevaluatedProperties": false
+/// })JSON");
+///
+/// sourcemeta::core::SchemaSchemaFrame frame;
+/// frame.analyse(document,
+///   sourcemeta::core::schema_official_walker,
+///   sourcemeta::core::schema_official_resolver);
+/// const auto result{sourcemeta::core::unevaluated(
+///     schema, frame,
+///     sourcemeta::core::schema_official_walker,
+///     sourcemeta::core::schema_official_resolver)};
+///
+/// assert(result.contains("#/unevaluatedProperties"));
+/// assert(!result.at("#/unevaluatedProperties").dynamic);
+/// assert(result.at("#/unevaluatedProperties").dependencies.empty());
+/// ```
+auto SOURCEMETA_CORE_JSONSCHEMA_EXPORT unevaluated(
+    const JSON &schema, const SchemaFrame &frame, const SchemaWalker &walker,
+    const SchemaResolver &resolver) -> SchemaUnevaluatedEntries;
 
 } // namespace sourcemeta::core
