+++ src/core/jsonschema/include/sourcemeta/core/jsonschema_resolver.h
@@ -6,6 +6,7 @@
 #endif
 
 #include <sourcemeta/core/json.h>
+#include <sourcemeta/core/jsonschema_types.h>
 
 #include <filesystem>  // std::filesystem
 #include <functional>  // std::function
@@ -14,29 +15,6 @@
 #include <string_view> // std::string_view
 
 namespace sourcemeta::core {
-
-// Take a URI and get back a schema
-/// @ingroup jsonschema
-///
-/// Some functions need to reference other schemas by their URIs. To accomplish
-/// this in a generic and flexible way, these functions take resolver functions
-/// as arguments, of the type sourcemeta::core::SchemaResolver.
-///
-/// For convenience, we provide the following default resolvers:
-///
-/// - sourcemeta::core::schema_official_resolver
-///
-/// You can implement resolvers to read from a local storage, to send HTTP
-/// requests, or anything your application might require. Unless your resolver
-/// is trivial, it is recommended to create a callable object that implements
-/// the function interface.
-using SchemaResolver = std::function<std::optional<JSON>(std::string_view)>;
-
-/// @ingroup jsonschema
-/// A default resolver that relies on built-in official schemas.
-SOURCEMETA_CORE_JSONSCHEMA_EXPORT
-auto schema_official_resolver(std::string_view identifier)
-    -> std::optional<sourcemeta::core::JSON>;
 
 /// @ingroup jsonschema
 /// This is a convenient helper for constructing schema resolvers at runtime.
